backend/simpleproject/requirements.txt
asgiref==3.8.1
Django==4.2.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
